Fix volume calculation for tiny polyhedra
This is a tricky fix, where the polyhedra is scaled up to a normalized
size, then scaled back down. For these tiny polyhedra, the volume is
on the order of machine epsilon, but the dimensions are several orders
of magnitude higher. Without scaling, erroneous negative volumes are
sometimes calculated. I'm not sure how legitimate this fix is, so it
might be better to just zero out these sufficiently small polyhedra.


git-svn-id: pececillo/branches/zjibben/flow/trunk@131