Make new polyhedron type
This changes the polyhedron type from containing itself (for nested convex polyhedra) to most of the
polyhedron functionality moving to the pure_polyhedron type, and the new polyhedron type containing
arrays of pure_polyhedron types.

The previous approach, with a polyhedron containing a pointer to other polyhedrons, seemed to be
unstable. Fortran should support this without issue, but it seemed to cause problems with automatic
destructors and deallocations.